+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+v0.3.1
+
+- Fixed issue when no config settings found. Added getProps() function
+  Error prevented anything from working. Changed to use getProps to prevent.
+
 v0.3.0
 
 - Fixed incorrect line endings. Updated front-end manifest.json. Fixed minor error in uibuilder.js. Updated dependency versions.